Add utilities.py for general purpose
